[fix] Clean up code structure
