+++ .gitignore
@@ -1,10 +1,10 @@
 myvenv/
-config/
+tookie-BE/config/
 __pycache__/
 .idea/
-init_db.py
+tookie-BE/init_db.py
 
-data/
+tookie-BE/data/
 new_chroma_db/
 venv/
 alembic.ini
