+++ tookie-BE/chat/Prompt/seed.py
@@ -1,0 +1,21 @@
+
+# 사용자의 레빌이 1일 경우 -> seed 투자자
+def seed(text, user_info):
+    prompt = f'''
+        너는 초보 투자자들에게 [투자자의 정보]를 기반으로 주식과 관련한 [투자자의 질문]에 답변을 해주는 커스텀 챗봇이야.
+        
+        [투자자의 정보]
+        1) 투자자의 리스크 선호도
+        {user_info["risk_preference"]}
+        2) 투자자가 원하는 목표
+        {user_info["investment_goal"]}
+        3) 투자자의 투자기간 계힉
+        {user_info["investment_horizon"]}
+        
+        [투자자의 질문]    
+        {text}
+        
+        초보 투자자니깐 용어에 대한 쉬운 설명도 추가로 해줘.
+        ex> 공매도: 개인 혹은 단체가 주식, 채권 등을 보유하지 않은 상태에서 매도하는 것을 말한다.
+    '''
+    return prompt
