+++ tookie-BE/user/settings.py
@@ -1,6 +1,4 @@
-import secrets
 from dotenv import load_dotenv
-import os
 from config.config import *
 
 # .env 파일에서 환경 변수 로드
